fix: tweaks

+++ packages/indexer/src/migrations/1704450403968_tokens-decimals.sql
@@ -1,6 +1,6 @@
 -- Up Migration
 
-ALTER TABLE "tokens" ADD COLUMN "decimals" INT;
+ALTER TABLE "tokens" ADD COLUMN "decimals" SMALLINT;
 
 -- Down Migration
